faet: find hydration solution!
