+++ src/modules/atomMap.ts
@@ -1,0 +1,8 @@
+import { testAtom } from './post/atom';
+
+const atomMapKey = testAtom.key;
+
+const atomMap = {};
+
+export { atomMapKey };
+export default atomMap;
