+++ src/pages/_app.page.tsx
@@ -1,13 +1,23 @@
-import type { AppProps } from "next/app";
-import theme from "@components/theme";
-import { ThemeProvider } from "styled-components";
-import GlobalStyle from "@components/GlobalStyle";
-import Head from "next/head";
-import { RecoilRoot } from "recoil";
+import type { AppProps } from 'next/app';
+import theme from '@components/theme';
+import { ThemeProvider } from 'styled-components';
+import GlobalStyle from '@components/GlobalStyle';
+import Head from 'next/head';
+import { RecoilRoot } from 'recoil';
+import { atomMapKey } from '@modules/atomMap';
+import { testAtom } from '@modules/post/atom';
 
 export default function App({ Component, pageProps }: AppProps) {
+  const { initialData: pageInitialData } = pageProps;
+  const initialRecoilState = {};
+  // console.log(pageInitialData);
+
   return (
-    <RecoilRoot>
+    <RecoilRoot
+      initializeState={({ set }) => {
+        set(testAtom, pageInitialData['posts']);
+      }}
+    >
       <ThemeProvider theme={theme}>
         <GlobalStyle />
         <Head>
